requirements.txt
Flask==2.0.1
Pillow==8.3.1
pytesseract==0.3.8
spacy==3.1.0
numpy==1.21.0